// generated/src/app/api-models/submitter-profile-detail.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class SubmitterProfileDetail {
  submitterId: string;
  fileType: string;
  dataType: string;
  mediaType: string;
  blockSize: number;
  customFormatPgm: string;
  encryptFormat: string;
  encryptKey: string;
  runOption1: string;
  runOption2: string;
  runOption3: string;
  runOption4: string;
  runOption5: string;
  comments: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  submitterDetailCode: string;
  description: string;
  userDefined1: string;
  userDefined2: string;
}